+++ gradle-extensions-subprojects/gradle-miscellanea-extensions/main/kotlin/com/javiersc/gradle/miscellanea/extensions/pluginArtifact.kt
@@ -1,0 +1,7 @@
+package com.javiersc.gradle.miscellanea.extensions
+
+import org.gradle.api.provider.Provider
+import org.gradle.plugin.use.PluginDependency
+
+public val Provider<PluginDependency>.pluginLibrary: Provider<String>
+    get() = map { "${it.pluginId}:${it.pluginId}.gradle.plugin:${it.version}" }